Cleanup the D2D1 generator constant buffer code
Also bumps the maximum constant buffer size

+++ src/ComputeSharp.D2D1/Interfaces/__Internals/ID2D1DispatchDataLoader.cs
@@ -14,5 +14,5 @@ public interface ID2D1DispatchDataLoader
     /// Loads the captured values used by the D2D1 shader to be dispatched.
     /// </summary>
     /// <param name="data">The sequence of serialized captured values.</param>
-    void LoadConstantBuffer(ReadOnlySpan<uint> data);
+    void LoadConstantBuffer(ReadOnlySpan<byte> data);
 }
